requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0
pydantic[email]==2.5.0
sib-api-v3-sdk==7.4.0
langchain==0.1.0
langchain-core==0.1.7
langchain-huggingface==0.0.3
langchain-groq==0.1.0
langgraph==0.0.26
transformers==4.36.0
torch>=2.4.0
huggingface-hub==0.19.4
python-dotenv==1.0.0
